mainGame/requirements.txt
pygame==2.5.2
websocket-client==1.6.4
requests==2.31.0
